+++ packages/ssx/src/__fixtures__/pages/blog/[slug].js
@@ -1,0 +1,7 @@
+import { html } from "@inglorious/web"
+
+export const blog = {
+  render() {
+    return html`<h1>Blog</h1>`
+  },
+}