+++ packages/ssx/src/__fixtures__/pages/index.js
@@ -1,0 +1,7 @@
+import { html } from "@inglorious/web"
+
+export const index = {
+  render() {
+    return html`<h1>Index</h1>`
+  },
+}