+++ packages/ssx/src/build.js
@@ -1,0 +1,13 @@
+// import { renderPage } from "./render.js"
+// import { getPages } from "./router.js"
+
+export async function build() {
+  // const pages = await getPages()
+  // for (const page of pages) {
+  // const module = await import(page.filePath)
+  // const { html, storeConfig, renderFn } = await renderPage(module, page)
+  // TODO: implement this
+  // Inject client script and write to dist/
+  // await writePageToDisk(page.path, html, { storeConfig, renderFn })
+  // }
+}