+++ examples/apps/web-logo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lit-todomvc
 
+## 0.1.6
+
+### Patch Changes
+
+- Add server-side router
+
 ## 0.1.5
 
 ### Patch Changes