+++ packages/ssx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/ssx
 
+## 0.2.0
+
+### Minor Changes
+
+- Add server-side router
+
 ## 0.1.4
 
 ### Patch Changes